Better extraction of uri password
