+++ src/test/java/com/ericsson/ei/mongo/MongoUriTest.java
@@ -14,7 +14,7 @@
    See the License for the specific language governing permissions and
    limitations under the License.
 */
-package com.ericsson.ei.utils;
+package com.ericsson.ei.mongo;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertFalse;
